Delete custom collector
