+++ build.gradle
@@ -49,7 +49,7 @@ run {
 
 application {
     mainModule = 'ip.counter.main'
-    mainClass = 'lv.id.jc.ipcounter.Main'
+    mainClass = 'com.yourcodereview.jegors.task1.Main'
 }
 
 jar {
