--- src/main/java/com/yourcodereview/jegors/task1/collector/package-info.java
@@ -1,4 +1,0 @@
-/**
- * The package contains the IP address collector.
- */
-package com.yourcodereview.jegors.task1.collector;
