+++ src/main/java/com/yourcodereview/jegors/task1/converter/IPv4Converter.java
@@ -3,7 +3,30 @@ package com.yourcodereview.jegors.task1.converter;
 import java.util.function.ToIntFunction;
 
 /**
- * Convert string representation of an IPv4 address to int.
+ * A simple converter to transform an IPv4 address from string representation to int.
  */
-public interface IPv4Converter extends ToIntFunction<CharSequence> {
+public class IPv4Converter implements ToIntFunction<CharSequence> {
+    private static final int DECIMAL_BASE = 10;
+
+    /**
+     * @param ipAddress representation of an IPv4 address
+     * @return int number representing the given address
+     */
+    @Override
+    public int applyAsInt(CharSequence ipAddress) {
+        long base = 0;
+        long part = 0;
+
+        for (int i = 0; i < ipAddress.length(); ++i) {
+            char symbol = ipAddress.charAt(i);
+            if (symbol == '.') {
+                base = (base << Byte.SIZE) | part;
+                part = 0;
+            } else {
+                part = part * DECIMAL_BASE + symbol - '0';
+            }
+        }
+        return (int) ((base << Byte.SIZE) | part);
+    }
+
 }
