+++ src/main/java/com/yourcodereview/jegors/task1/converter/package-info.java
@@ -3,7 +3,7 @@
  * <p>
  * There is one simple implementations of the interface in the package:
  * <p>
- * {@link com.yourcodereview.jegors.task1.converter.IPv4SimpleConverter}
+ * {@link com.yourcodereview.jegors.task1.converter.IPv4Converter}
  * This converter converts the representation of an IPv4 address from a string type
  * to a number. The converter does not check the string representation of the ip address.
  * In the case of an erroneous address format, the returned number is unpredictable.
